+++ CHANGELOG.md
@@ -6,4 +6,5 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
-- Initial release
+- Implement graphical CPU percentage history
+- Implement CPU percentage meter